Tic Tac Toe  - Game
Tic Tac Toe is a popular game that can be created using Python programming language. The game is played on a 3x3 grid, and the players take turns marking either X or O on the grid. The first player to get three of their marks in a row (either horizontally, vertically, or diagonally) wins the game. If all the spaces are filled and no player has three marks in a row, the game ends in a tie.

To create the Tic Tac Toe game using Python, we first need to create a 3x3 grid that will serve as the game board. We can represent the grid using a list of lists, where each element in the list represents a row on the game board.

Next, we need to create a function that will allow the players to take turns marking X or O on the game board. We can do this by asking the players for their move, and then checking if the move is valid (i.e. if the space is empty). If the move is valid, we can mark the space with the player's symbol (X or O). If the move is not valid, we can ask the player to try again.

We also need to create a function that will check if there is a winner. We can do this by checking each row, column, and diagonal on the game board to see if all the spaces are marked with the same symbol (X or O). If there is a winner, we can display a message and end the game.

Finally, we need to create a function that will check if the game has ended in a tie. We can do this by checking if all the spaces on the game board are filled and if there is no winner. If this is the case, we can display a message and end the game.

Overall, creating Tic Tac Toe using Python is a fun and rewarding project that can help beginners learn the basics of programming. With just a few lines of code, we can create a fully functional game that can provide hours of entertainment.